+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [0.1.3+snapshot25w24a] - 2025-06-08
+
+### Changed
+
+- Changed wordsOnStream to wordleonwheels
+- Moved Config file stuff to files.py
+- Moved time code to make it more accurate
+- Now using "%" formatting in logging functions: [logging-fstring-interpolation / W1203](https://pylint.readthedocs.io/en/latest/user_guide/messages/warning/logging-fstring-interpolation.html)
+- Put the bot's code in twitch_bot.py
+- Changed c!online to c!status
+- Moved logging functionality to logger.py
+- Automatically stops searching for word when a user correctly guess the word.
+
+### Added
+
+- Added c!mud_curses command
+- Added more comments and docstrings
+- Added a way to get word into the program :)
+
+### Removed
+
+- Removed 'word' field in config.toml
+
 ## [0.1.2+snapshot25w23b] - 2025-06-07
 
 ### Added
@@ -43,3 +66,4 @@ _Initial release_
 [0.1.0]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.0
 [0.1.1+snapshot25w23a]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.1+snapshot25w23a
 [0.1.2+snapshot25w23b]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.2+snapshot25w23b
+[0.1.3+snapshot25w24a]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v0.1.2+snapshot25w24a